adds seperate hero component for fixtures
